# Convert TiledColor To Struct