… src/DotTiled/Properties/TiledColor.cs