Converts `TiledColor` to a struct similar to how `Color` is defined in `System.Drawing` and other libraries. The actual value is a 32-bit unsigned integer with each octet representing a different component. Also defines several standard static colors that can be used.

These changes does manage to reduce the allocated memory and increases performance by a smidge.