@7H3LaughingMan 2025-05-01
```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.26100.3775)
Intel Core i5-14500, 1 CPU, 20 logical and 14 physical cores
.NET SDK 9.0.203
  [Host]     : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.15 (8.0.1525.16413), X64 RyuJIT AVX2
```

### Before Changes

| Method      | Categories               | Mean     | Median   | Ratio | Gen0   | Gen1   | Allocated | Alloc Ratio |
|------------ |------------------------- |---------:|---------:|------:|-------:|-------:|----------:|------------:|
| DotTiled    | MapFromInMemoryTmjString | 4.953 μs | 4.937 μs |  1.00 | 0.4730 |      - |   5.81 KB |        1.00 |
| TiledLib    | MapFromInMemoryTmjString | 6.311 μs | 6.197 μs |  1.30 | 0.7401 | 0.0153 |   9.07 KB |        1.56 |
|             |                          |          |          |       |        |        |           |             |
| DotTiled    | MapFromInMemoryTmxString | 3.129 μs | 3.116 μs |  1.00 | 1.3580 | 0.0610 |  16.72 KB |        1.00 |
| TiledLib    | MapFromInMemoryTmxString | 5.533 μs | 5.549 μs |  1.77 | 1.8921 | 0.1221 |  23.31 KB |        1.39 |
| TiledCSPlus | MapFromInMemoryTmxString | 6.857 μs | 6.819 μs |  2.19 | 2.6855 | 0.1221 |  33.23 KB |        1.99 |

### After Changes

| Method      | Categories               | Mean     | Median   | Ratio | Gen0   | Gen1   | Allocated | Alloc Ratio |
|------------ |------------------------- |---------:|---------:|------:|-------:|-------:|----------:|------------:|
| DotTiled    | MapFromInMemoryTmjString | 4.428 μs | 4.419 μs |  1.00 | 0.4425 |      - |   5.51 KB |        1.00 |
| TiledLib    | MapFromInMemoryTmjString | 6.252 μs | 6.224 μs |  1.41 | 0.7401 | 0.0153 |   9.07 KB |        1.65 |
|             |                          |          |          |       |        |        |           |             |
| DotTiled    | MapFromInMemoryTmxString | 3.042 μs | 3.036 μs |  1.00 | 1.3390 | 0.0610 |  16.41 KB |        1.00 |
| TiledLib    | MapFromInMemoryTmxString | 5.801 μs | 5.710 μs |  2.00 | 1.8921 | 0.1221 |  23.31 KB |        1.42 |
| TiledCSPlus | MapFromInMemoryTmxString | 6.755 μs | 6.684 μs |  2.23 | 2.6855 | 0.1831 |  33.16 KB |        2.02 |